Begin refactoring SRA tests out and to use realWin
